♻️ refactor: champ -> champion 으로 줄임말 제거
