--- src/renderer/src/features/champ/ChampDetail/index.ts
@@ -1,3 +1,0 @@
-// === Automatically generated file. Don't edit it. ===
-export * from './ChampDetail'
-export { default } from './ChampDetail'
