+++ src/renderer/src/features/champion/ChampionStats/index.ts
@@ -1,0 +1,3 @@
+// === Automatically generated file. Don't edit it. ===
+export * from './ChampionStats'
+export { default } from './ChampionStats'
