+++ src/renderer/src/features/inGame/InGameInfo/InGameInfo.tsx
@@ -57,7 +57,7 @@ const InGameInfo = ({ className }: InGameInfoProps) => {
   }
 
   const moveChampionDetailPage = (championId: number) => {
-    navigate(`/champ/${championId}`)
+    navigate(`/champion/${championId}`)
   }
 
   return (
