+++ src/renderer/src/features/tier/TierTable/TierTable.tsx
@@ -90,7 +90,7 @@ const TierTable = ({ className }: TierTableProps) => {
       <Table
         onRow={record => ({
           onClick: () => {
-            navigate(`/champ/${record.championId}?laneId=${laneId}`)
+            navigate(`/champion/${record.championId}?laneId=${laneId}`)
           },
         })}
         components={{
